requirement.txt
click==8.1.3
colorama==0.4.5
Flask==2.2.2
Flask-SQLAlchemy==3.0.1
greenlet==1.1.3.post0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
PyJWT==1.7.1
SQLAlchemy==1.4.41
Werkzeug==2.2.2
